moved DisjointSet into cluster.hierarchy added toctree entry
